Separate investigation and environment classes while avoiding circular imports
